DevTools: [HeapTracking] fix for selected size calculation.
We print selected size at status bar. Everything works fine except the ranges
which include the leftmost item. Actually we don't draw the bar for it because it is huge.
But when the leftmost item is selected then we show in the grid all the objects which were found in the heap at start tracking moment.
At the same time the size in the status bar doesn't reflect this fact.

BUG=none
TEST=none
R=apavlov@chromium.org

Review URL: https://codereview.chromium.org/15736006

git-svn-id: svn://svn.chromium.org/blink/trunk@150900 bbb929c8-8fbe-4397-9dbb-9b2b20218538