DevTools: Do not automatically enable Canvas Profiler upon switching to the Profiles tab (2nd try).
2nd try after rollback: https://src.chromium.org/viewvc/blink?revision=151414&view=revision
Addressed the crash issue, fixed a canvas profiler test.

Enabled Canvas Profiler may be intrusive for other profilers, thus we should not automatically enable it.
Instead we will provide a Enable/Disable button which would do the action and reload the debuggee page if necessary.
Also, if Canvas Profiler option is selected but it's not (yet) enabled, we should disable the "Take Snapshot" button.

BUG=236767
R=pfeldman@chromium.org

Review URL: https://codereview.chromium.org/16189013

git-svn-id: svn://svn.chromium.org/blink/trunk@151534 bbb929c8-8fbe-4397-9dbb-9b2b20218538