DevTools: update scoring function for "Open File" dialog
This change updates scoring function which is used in Ctrl-O dialog.
The improvement turns out to be up to 3x times faster then the current
implementation on the requests I tested ("froscr", "dte", "cmte", "insins",
"adscon"). The top match for the given requests are the same (or even better)
compared to the current implementation. Moreover, the algorithm respects
the CAPS in the query.

Few words about the general approach implemented in the patch.
Given the query Q and data entry E, we count a dynamic
score[i][j] which represents the best score archieved for the Q's prefix of
length i and E's prefix of length j. The recurrental formulae is

score[i][j] = MAX(score[i-1, j-1] + match(Q[i], E[j]), score[i, j - 1])

where 'match' function is a weighting function.
Additionally, the algorithm constructs a matrix 'sequence' which holds
amount of consequtive chars for the score[i][j]; this matrix is passed to
the weighting function 'match' and it is also used for answer backtracking.

BUG=

Review URL: https://chromiumcodereview.appspot.com/17583009

git-svn-id: svn://svn.chromium.org/blink/trunk@153137 bbb929c8-8fbe-4397-9dbb-9b2b20218538