DevTools: Popover sometimes closes even when the mouse moves inside.
This happens with ObjectPopoverHelper, where popover's contentElement is a SPAN with "white-space: pre".
In this case mousemove is not being caught if the mouse moves over an empty space inside the popover, not targeting the child contentElement SPAN, and popover closes after a timeout.

Thus we should listen mousemove on the parent DIV element (this._contentDiv).
We should NOT listen mousemove on this.element because of it's paddings, in which case the popover would not close when the mouse is outside of the visible area, but over the padding space (this was tried before: https://bugs.webkit.org/show_bug.cgi?id=105167).

BUG=274263
R= eustas@chromium.org, vsevik@chromium.org, pfeldman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22906006

git-svn-id: svn://svn.chromium.org/blink/trunk@156223 bbb929c8-8fbe-4397-9dbb-9b2b20218538