DevTools: Show object properties in quotes when needed in console preview.
If an object property contains whitespaces (or some other chars) we should enclose it in quotes in console preview.

We use a simplified regexp to test if a name might be an IdentifierName, and if so we do NOT use quotes (see the docs below).

Object Initialiser syntax: http://ecma-international.org/ecma-262/5.1/#sec-11.1.5

BUG=245642
R=pfeldman, vsevik

Review URL: https://chromiumcodereview.appspot.com/24479003

git-svn-id: svn://svn.chromium.org/blink/trunk@158439 bbb929c8-8fbe-4397-9dbb-9b2b20218538