DevTools: Timeline: remove TimelinePresentationModel.Category dependency from TimelineOverviewPane
TOP has nothing in common with TPM.Categories. So we can move that code
to the TimelineEventsOverview which is only affected class.

TOP created Overviews each time when user selected other presentation.
In combination with listeners in TimelineEventsOverview that would be a reason of a memory leak.
Also it was unnecessary because we have only one overview of each type.

BUG=none
R=pfeldman@chromium.org

Review URL: https://codereview.chromium.org/111453002

git-svn-id: svn://svn.chromium.org/blink/trunk@163526 bbb929c8-8fbe-4397-9dbb-9b2b20218538