Timeline: disable start/stop button and properly indicate progress when stop is pending
Now that we may fetch a lot of trace (or buffered timeline) events after
timeline is stopped, we need a better indication of that. The button remaining
red after it's pressed is very annoying. Let's turn it off and disable till
we get all the events, and if the progress pane is on, display the appropriate
message there.

BUG=361045

Review URL: https://codereview.chromium.org/322403004

git-svn-id: svn://svn.chromium.org/blink/trunk@176214 bbb929c8-8fbe-4397-9dbb-9b2b20218538