DevTools: [SSP] fix "insert style rule" button on mac
This patch moves insert style rule button further to the left to not overlap with scrollbar. It also fixes its width.

As a drive-by, this patch sets "translateZ" rule hack on styles sidebar pane as a workaround for compositor bug. This makes scrollbar appear on MB Air.

BUG=404102
R=pfeldman

Review URL: https://codereview.chromium.org/477113002

git-svn-id: svn://svn.chromium.org/blink/trunk@180359 bbb929c8-8fbe-4397-9dbb-9b2b20218538