Move rendering/RenderLayer* to layout/
This CL moves the various rendering/RenderLayer* files into layout/. In most
cases the files were renamed to be Layer*. The one exception is
RenderLayerModelObject was renamed to be LayoutLayerModelObject to signify it
is in the Layout tree.

This CL updates RenderLayerCompositor to be LayerCompositor to match this name
change.

BUG=450612
TBR=jchaffraix@chromium.org, mkwst@chromium.org

Review URL: https://codereview.chromium.org/898783003

git-svn-id: svn://svn.chromium.org/blink/trunk@189529 bbb929c8-8fbe-4397-9dbb-9b2b20218538