Rename {Author,UserAgent}ShadowRoot to {Open,Closed}ShadowRoot.
This batch-renaming is the very first step to support a closed shadow tree. No functional changes at all.

The relevant W3C spec issues:
- [Shadow]: Define the behavior of *closed* shadow trees
  https://www.w3.org/Bugs/Public/show_bug.cgi?id=27775
- [Shadow]: Add "closed" flag to createShadowRoot
  https://www.w3.org/Bugs/Public/show_bug.cgi?id=20144

See the bug 459136 for the details of the motivation.

The file names of existing layout tests aren't touched at all in this CL so that we can isolate an issue.

BUG=459136

Review URL: https://codereview.chromium.org/935283002

git-svn-id: svn://svn.chromium.org/blink/trunk@190536 bbb929c8-8fbe-4397-9dbb-9b2b20218538