DevTools: clear flame chart when selected range is empty in frame overview
When selected range on the Timeline frame overview doesn't contain any actual frame we treat that as if the selection was empty. There is no reasonable time scale to draw on the flame chart in this case so we simply clear the whole chart along with time dividers.

BUG=487162

Review URL: https://codereview.chromium.org/1144673004

git-svn-id: svn://svn.chromium.org/blink/trunk@195561 bbb929c8-8fbe-4397-9dbb-9b2b20218538