Devtools: Allow ToolbarItem to have a defined action
This change allows for explicitly defined actions for a ToolbarItem. Toolbar
buttons and counters trigger this action when the item is clicked.

For ExtensibleToolbar buttons, the declarative action id is captured.
This change makes way for tooltips to be able to display keyboard shortcuts
associated with the toolbar item's action.

BUG=513066

Review URL: https://codereview.chromium.org/1243973002

git-svn-id: svn://svn.chromium.org/blink/trunk@199343 bbb929c8-8fbe-4397-9dbb-9b2b20218538