Devtools UI: Add node-specific actions into a ghost toolbar in DOM
This change adds a button in the elements gutter which toggles the node-specific toolbar. This toolbar is triggered from this icon or markers when present. When markers are present, the icon is not visible and the toolbar is automatically opened.

This change also removes the elements toolbar completely by moving layout mode to the styles sidebar pane.

BUG=528387

Review URL: https://codereview.chromium.org/1304173004

git-svn-id: svn://svn.chromium.org/blink/trunk@202482 bbb929c8-8fbe-4397-9dbb-9b2b20218538