lib: re-implement web-socket server using a core.async channel as queue
Previous implementation has been calling http/on-receive too late.
It could happen that client sends a message too quickly after opening
a connection that we could sometimes miss it.

New approach registers http/on-close and http/on-receive ASAP. But
instead of processing messages right away it is putting them on
core.async channel. When we are ready after accepting client connection
and have all initialisation set up, we can start processing
jobs on the channel asynchronously.