Send certificates to devtools when it's open instead of using serialized SSLStatus and sending certificate IDs.
This is along the earlier patchsets of stopping to send SSLStatus to the renderer. This will help network service (since currently it's what sends this data, and it shouldn't know about UI state) and PlzNavigate (since this is a precursor for removing cert_ids which depend on having a renderer process during navigate).

Since this patch has gotten large, I'll remove sending sending security_info to the renderer in a followup patch.

BUG=598073, 504347

Review-Url: https://codereview.chromium.org/2296953004
Cr-Commit-Position: refs/heads/master@{#416700}