Don't trigger Object.prototype setters when generating completions
Autocomplete gets its candidates for completion by running a function in
the context of the page. If we use a plain object to hold results, we'll
end up triggering any setters that the page added to Object.prototype.

Using a prototype-less object avoids this, but a page could still
interfere by replacing Object.create (or Object.getPrototypeOf, String,
Number, Boolean...).

BUG=645328

Review-Url: https://codereview.chromium.org/2324093002
Cr-Commit-Position: refs/heads/master@{#417729}