tests: teach browser task runner to attempt to kill derailed tasks
The issue is that "frozen" tasks can a devtools instance open
and that instance is using chrome debugging port. Without teardown
of this connection we likely fail when moving onto a next task.

This new implementation tries to ask task runner to kill current
running task and perform proper teardown. After that we should be
safe running next task.