DevTools: eliminate ToolbarButton.setState method; cleanup toolbar.css
This patch is a preparation for the crrev.com/2473313002. It cleans up
toolbar.css so that for every icon in toolbarButtonGlyphs.png there's
exactly one css class which defines it.

Today, some clients of the ToolbarButton were abusing ToolbarButton
state to specify different icons. This patch moves these clients to
use ToolbarButton.setGlyph method.

As a result, the ToolbarButton.setState method became unneeded.
Instead, there's a ToolbarToggle.setToggleWithRed method which is there
to support needs of "recording" red icon.

As a nifty perk, the patch gets rid of two icons in
toolbarButtonGlyphs.svg.

BUG=none
R=dgozman

Review-Url: https://codereview.chromium.org/2482593003
Cr-Commit-Position: refs/heads/master@{#430714}