Do not prompt for reload when security panel is opened on an interstitial
Because interstitials are not real navigations, the security panel
sidebar doesn't work properly for an interstitial, and also doesn't
provide that much more useful information than what's shown in the
Overview view. For this reason, we were already hiding the origins
sidebar when you navigate to an interstitial while the security panel is
open. This CL extends that behavior to when you open DevTools while an
interstitial is already showing.

As a drive-by, this CL also fixes a bug where requests that finished after
the interstitial was shown would still show up in the sidebar.

Also note that this CL only partially addresses the linked bug. This CL
removes the confusing/useless prompt to reload, but doesn't fix the fact
that refreshing on an interstitial does weird things (sends you back to
the previous page). The latter weirdness has been split into
https://crbug.com/669316.

BUG=637299
TEST=Navigate to https://expired.badssl.com. Open the DevTools security
panel. Observe that there is no prompt to reload in the sidebar.

Review-Url: https://codereview.chromium.org/2522733002
Cr-Commit-Position: refs/heads/master@{#434887}