DevTools: fix search across all files in case of "file:" prefix
Search across files supports narrowing of the search scope with the
"file:" pattern. For example:

   [1]    file:front_end QUERY
   [2]    QUERY file:front_end

The bug happens only in the [1] case, when the query is prefixed with
"file:" pattern. In this case, we currently consider the query
to be " foobar" instead of "foobar" (e.g. containing excessive space in
front).

This patch swallows all the whitespaces around the "file:" pattern so
that they are not included in the actual search queries.

BUG=686359
R=dgozman

Review-Url: https://codereview.chromium.org/2660853002
Cr-Commit-Position: refs/heads/master@{#447196}