[Devtools] Fixed flicker in network for large stream of requests
This patch updates network to update the ViewportDataGrid instantly
which fixes the flicker.

Reason this occurs is because network gets an update to a request which
schedules an update. It finds a new node now needs to be added so it
sends it to datagrid and waterfall to be added which Waterfall processes
immidatly, but ViewportDataGrid schedules the next frame to insert the
new nodes. The scroller then gets synced and stick to bottom kicks in
which causes the scroller to scroll down which reveals blank area in the
ViewportDataGrid.

R=pfeldman
BUG=697273

Review-Url: https://codereview.chromium.org/2726763002
Cr-Commit-Position: refs/heads/master@{#454319}