DevTools: fix stretch layout in search/replace bar
SearchableView used to build the find and replace UI using <table>. A table
inside a flex container produced the strange behavior in the bug.

This CL fixes the visual bug, removes the table elements, and changes the layout
of the find/replace text inputs to stretch as much as possible.

BUG=698830

Review-Url: https://codereview.chromium.org/2734043002
Cr-Commit-Position: refs/heads/master@{#455244}