DevTools: clicking in console messages should not jump to bottom
ConsolePrompt's moveCaretToEndOfPrompt() works by setting the CodeMirror
editor's selection to the end. After a recent CodeMirror roll, doing so now
scrolls the editor into view.

This change in behavior led to an error where expanding an object in console
would jump the viewport to bottom. This CL only focuses the prompt without
moving the cursor, in some cases.

BUG=706128

Review-Url: https://codereview.chromium.org/2840663002
Cr-Commit-Position: refs/heads/master@{#468009}