[DevTools] Enable network capability in worker unconditionally
We are planning to start the field trial experiment of
off-main-thread-fetch optimization.

But in the current implementation, web developers need to enable
'Network in workers' experiment of DevTools to show the network requests
from workers in DevTools if the optimization is enabled in their Chrome.
It is very troublesome for them while debugging PWA.

To avoid this problem, this CL enables the network capability in worker
unconditionally.

Bug: 443374
Change-Id: I5ed24996cd0ff30a5bd22ce1067f653a5c68ae9b
Reviewed-on: https://chromium-review.googlesource.com/595294
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#491559}