DevTools: test path for experiment in new test framework
The legacy test framework has a mechanism for enabling experiments at the startup
of DevTools, and this brings parity in the new test framework.

In the long-term, there should be a better solution as this is very
confusing for first-time DevTools test contributors and is a code smell.

Bug: 667560
Change-Id: I230453e16b3d06f6ee9ada99c0beb793f244d8c5
Reviewed-on: https://chromium-review.googlesource.com/695910
Reviewed-by: Will Chen <chenwilliam@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Will Chen <chenwilliam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#505901}