DevTools: fix geolocation toSetting() serialization
The Geolocation struct used to hold a string-type error property,
but was changed into a boolean here:
https://codereview.chromium.org/1917543002

There was a mistake in that CL, where it was still expecting a
string-type error, causing serialization to always return an
empty string.

This CL fixes the serializer, allowing the last applied
geolocation to be saved.

Bug: 649657
Change-Id: I7f65f7844dca9ee2e182970602c6810a3c776973
Reviewed-on: https://chromium-review.googlesource.com/778044
Commit-Queue: Erik Luo <luoe@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#519953}