DevTools: refreshing Elements search should not show '0 out of 0'
The Elements search bar shows the number of matches when active.
Upon a refresh/navigation, it cancels the search, but leaves
'0 out of 0' in the search box, even if there are matches.

This CL stops showing incorrect information and hides the match
count completely in those cases.

Screenshot: https://imgur.com/a/xdyD4

Bug: 554654
Change-Id: I49791d302ade2719ae0b609cab8d325a562bb87a
Reviewed-on: https://chromium-review.googlesource.com/920041
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Erik Luo <luoe@chromium.org>
Cr-Commit-Position: refs/heads/master@{#537847}