Developer Tools/ObjectUI - less options for synthetics.
Synthetic properties (like "Return value") do not have
a meaningful property path, so the option to copy it is
useless, so is the option to add it to watch.

Internal properties (like "[[PromiseValue]]" and so on)
are also synthetic (in terms of not being accessible
from JavaScript), so now they are marked as such.

Bug: none
Change-Id: I381098e478f0217a835b9d8b5ca5f3ba1304da3c
Reviewed-on: https://chromium-review.googlesource.com/1084093
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: PhistucK <phistuck@gmail.com>
Cr-Commit-Position: refs/heads/master@{#570499}