[DevTools] fix(elements): preserve active text highlight on contextmenu
Previously we would listen to a blur event coming out of the HTML editor
to know when to commit the text. To prevent an internal focus change
we would cancel mousedown events inside the editor. This confused
codemirror, which expects focus to have moved after a right click
and messed up its text selection color.

This patch removes the mousedown canceling, and checks relatedTarget
on the focus event to only commit when focus leaves the editor.

Bug: 894372
Change-Id: Id29f26c75d9f5a12c52679c3cd947987d2b8779e
Reviewed-on: https://chromium-review.googlesource.com/c/1277810
Commit-Queue: Joel Einbinder <einbinder@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#599015}