DevTools: Support time-based overview for sampling profiles.
The patch adds timeline overview into sampling profiles types.
During sampling recording it periodically requests current samples
and displays them on the overview with the bar chart.

After recording is complete it is possible to select a range and
check allocations stacks attributed to the range.

The new mode is put behind an experiment.

BUG=889545

Change-Id: I186a1869ee9e99d9431260e728ce122f38b1ea99
Reviewed-on: https://chromium-review.googlesource.com/c/1297280
Commit-Queue: Alexei Filippov <alph@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#602872}