[DevTools] Handle long string in service worker list's script name and clients
If the name string of the service worker script or a service worker
client is too long, this patch restricts its width and denote the
overflow with ellipsis. Users can still get the string by hovering
above it, and the string is still clickable.

Bug: 847801
Change-Id: I9d3a07c410af6bb98ae42046894ea3d7223afcbc
Reviewed-on: https://chromium-review.googlesource.com/c/1312973
Commit-Queue: Haihong Li (Harley) <hhli@chromium.org>
Reviewed-by: Erik Luo <luoe@chromium.org>
Cr-Commit-Position: refs/heads/master@{#604760}