Show the correct transfer size of signed exchange in DevTools
Currently DevTools show the size of HTTP header of signed exchange response.
This is confusing for developers.

This CL will fix this issue by setting the encoded_data_length in
URLLoaderCompletionStatus of the downstream from the network received in
SignedExchangeLoader, which should correspond to that of the inner response, and
setting it to the transfer size of the outer response in NetworkManager.js.

Bug: 928137
Change-Id: Id075ad95aa187002866ba174a01b6ae400e2ac05
Reviewed-on: https://chromium-review.googlesource.com/c/1449958
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#629022}