[DevTools] Add request headers to cached response matching
In Applocation>Cache Storage, to preview a cached response, DevTools
needs to call requestCachedResponse() to the backend to fetch it.
Originally, DevTools is unable to fetch cached responses for requests
that have headers, and this is because in DevTools's beckend, the field
'headers' of mojom::blink::FetchAPIRequest remains unfilled. This
patch fixes the issue by filling this field.


Bug: 890359
Change-Id: I599cc1393d5c02baa7d7f90fc22fd379b1745090
Reviewed-on: https://chromium-review.googlesource.com/c/1455272
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Haihong Li (Harley) <hhli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#630005}