DevTools: Order split widget children correctly in the DOM
Previously, the sidebar and main widgets were arranged via css. This
caused the tab order to be wrong for keyboard navigation.

This patch puts the dom elements in the correct order, which lets us
remove the CSS hacks and fixes the tab order.

If the main widget and the sidebar widget both contain extension
iframes, we can't swap the widgets without destroying the extensions.
This is quite an exotic scenario though, and currently never happens
in DevTools.

Bug: 941182
Change-Id: I3a93d7f3e6fb52f5f7c1b1a8511f713d31ae07ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1521798
Reviewed-by: Erik Luo <luoe@chromium.org>
Commit-Queue: Joel Einbinder <einbinder@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640657}