DevTools: cleanup keydown pattern
We sprinkle a lot of isEnter/isSpace keychecks in the code for invoking
buttons and links and it has started to smell.

This change adds a new helper isEnterOrSpaceKey() to save some typing.
It also adds a new onInvokeElement that conditionally invokes a callback
when a keydown event that passes isEnterOrSpaceKey occurs.

Change-Id: I163ad059db6d58da8209ed6887b3d728deb97cec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716096
Commit-Queue: Jeff Fisher <jeffish@microsoft.com>
Reviewed-by: Joel Einbinder <einbinder@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686753}