DevTools: Settings/Devices/Add/Cancel triggers validation, fails cancelling
This bug is caused by the fact that the "cancel" button is moved as a
result of validation, causing the 'click' event to not fire. Per code
review feedback, the change I've made is to reposition the error message
to be below the commit/cancel buttons. The result of that is to very briefly
see the validation error before the form dismisses.

Before and after images: https://imgur.com/a/7LseDbs

Note: Appears to also address bug 986776.

Bug: 983003, 986776
Change-Id: I8d753faf59c1b6967639d80cedbfe8cce5451448
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716329
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#698308}