DevTools: Hide not-found sourcemap refs
When a JS source file refers to a sourcemap, which then refers to a file
that is not found on the file system (file://), the underlying request
handler doesn't return any headers. With no headers, the DevTools UI
bindings component presumes that the response is equivalent to a 200 OK
and returns a blank file. Instead, we're going to return a 404, which
will then prevent DevTools from surfacing the not-found file altogether.

Bug: 941035
Change-Id: Icf92bdcfb415f0381217547dac9037d53915e3fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847833
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#704396}