DevTools: a11y of Elements/Accessibility Computed Properties pane
This change enables keyboard selection and navigation, as well as
screen reader, for the Computed Properties pane of the Accessibility
tool. Related nodes can now also be navigated-to by pressing the
<enter> key while these nodes are selected.

This change is a break-out and continuation of CL 1603344.

Screencap: https://imgur.com/a/tcipWNX
Change-Id: Id7702aa8c2d32cc0afe5215006db7f16b80af195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1812105
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#704869}