[devtools] Report addition of new files as coverage update
We report the <script>function foo() {}</script> as completely
uncovered, and this caused the CoverageModel to determine that the
coverage report from the back-end did not change anything (used bytes
did not increase). However, the coverage report from the back-end might
have added the script to the CoverageModel's entries in the first place.
In such cases, the coverage list view was not properly updated.

This CL adds an update notification whenever a CoverageURLInfo is first
created as well (independent of whether used bytes are zero).

Bug: chromium:1004203
Change-Id: I458e14725a10a25087e443b2c509dc444d4bf997
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864675
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706363}