[devtools] Add find next with Ctrl+G and F3
This cl adds Ctrl+G and F3 as keyboard shortcuts
for windows and linux to find the next occurrence
while a search is open (not necessarily focused).
The default behaviour of Ctrl+G opening the goto
line popup still works as long as the search is not
opened.

Bug: chromium:887586
Change-Id: I8164786778e0b8cb99f7d6ca9d37a61cc34b622b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1817887
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706828}