DevTools: Race condition crash on HTML edit/resize
In the Elements tool, there is a race condition that can occur
between the deletion of the HTML editor component and the resize
logic, the latter of which unconditionally accesses the HTML
editor without checking to see whether it's present.

This change adds a declaration of the HTML editor element as a
member of Elements.ElementsTreeElement (so that Closure will
typecheck it), and guards against this particular error from
occurring.

Bug: 1016391
Change-Id: I323fe5a14caebcc0dd05546045a0b08d9c5dc604
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1872387
Commit-Queue: John Emau <John.Emau@microsoft.com>
Reviewed-by: John Emau <John.Emau@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>