DevTools: Hide datagrid from SR when landing page is open
Previously, when navigating the Network landing page with a screen
reader in browse/scan mode, the user would be able to reach the
datagrid elements that are not visible. This content is reached before
reaching the landing page, so a user wouldn't know to keep navigating
past the grid to reach the shortcut suggestions.

This change adds the hidden class to the datagrid and toolbar which
sets display: none !important for these elements. This hides the data
grid and summary toolbar from screen readers so focus moves straight to
the landing page content.

Bug: 963183
Change-Id: Ie767f8c9933675a6bc6b16a50f2ec555f08ef5a6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1880306
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Commit-Queue: Amanda Baker <ambake@microsoft.com>