DevTools: Make request headers keyboard navigable
Currently, the request headers tab cannot be manipulated with a
keyboard whatsoever. This means that a keyboard user cannot expand or
collapse the groups in the tree to focus on the content that they want
to see.

This change enables the keyboarding by allowing selection of the tree
and its nodes, by fixing the focus styling (so focus is always visible),
and by selecting the first node as an entry point for tabbing into the
content.

Screenshot: https://imgur.com/a/8H8Xu0z

Bug: 963183
Change-Id: Ie48032436ef91b5d7de974a728f184021e20de1e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1880311
Commit-Queue: Amanda Baker <ambake@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>