DevTools: Add context menu entry for "Show more" button
In the Network Headers pane long Request Payloads are truncated and have
a "Show more" button that allows a user to show the entire payload. This
button is not keyboard accessible.

This change adds a context menu entry to the tree item containing the
request payload, so that a keyboard user can access this action.

Testing: use bing.com and filter network requests for "xls.aspx". The
button will be in the Request Payload

Screenshot: https://imgur.com/a/DX7ZlOr

Bug: 963183
Change-Id: Id7e9fec6861be5627eff1007cd27643cf9c426f2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1880312
Commit-Queue: Amanda Baker <ambake@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>