Network A11y: Improve keyboarding experience of Network tool Datagrid
This changes builds off of
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1881219
in order to improve the keyboard accessibility
of the Network tool Datagrid.

Previously, the network tool ties selection of a node in the grid
directly to opening the request sidepane. This makes it challenging for
keyboard only users to view and navigate the network tool datagrid
columns.

This change separates selection from opening the sidepane, and fixes
styling issues that affect the columns that were previously hidden
behind the sidepane when a node is selected. Selected nodes are
highlighted in the grid, and are "activated" using the enter key
or a click in the primary column to open the sidepane.

Before: https://imgur.com/U1PG6Hr

After: https://imgur.com/eXgDE9D

Bug: 963183

Change-Id: I2c2629670624e807adb4cd9e9e9b75cdafbae7a5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1881592
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>