DevTools: Fix non-dismiss of request inspector on reset
When a user toggles off and then on the network monitoring, without the
Preserve Log option enabled, the network log resets. However, if the
individual request inspection pane is open, it will stay open. This is due
to a parameter in NetworkLogView#_reset being true instead of false.

However, correcting this behavior has a side effect that, when enabling
the logging option in the same manner as above but with the keyboard, it
shifts focus to the network data grid (which is empty). This causes an
a11y problem because NVDA just announces "blank". So this change removes
the call to NetworkLogView#resetFocus from the handler within
NetworkPanel#_hideRequestFocus().

However, that change ALSO causes a problem, in that pressing <Esc> uses
that code path. Instead, we'll call NetworkLogView#resetFocus from the
keyboard handler which is the NetworkPanel action delegate.

Bug: 1025332
Change-Id: Ida7b78447237914df5f51b31edc7ab8cf2aa7c56
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1918734
Reviewed-by: Mike Jackson <mjackson@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>