Do not draw coalesced latency info in performance panel
Display events that were coalesced into others in performance tracing
are not useful and confusing. This CL make it filter out these events.

(Maybe it'd be nice to also mark coalesced events in the slices, but
maybe it'd make the ui too complicate, and the info not very useful to
web developers)

Bug: 956681
Change-Id: I4499cae055f305ebc247242602a95dfcf224f784
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1928916
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>