Do not load ARIAProperties as an ES module
ARIAProperties is a script downloaded from the appspot server.
Therefore, it should not be imported using an ES import, but be a script
that is included in the accessiblity_module.js instead.

A locally built Chromium version does now properly show the
Accessibility pane in the Elements panel. However, it does not include
the ARIAProperties definitions yet, as we will need to wait until these
files are on the appspot server.

We need to check the Canary version tomorrow to verify that the file is
correctly included and properly loaded using the accessibility_module.js

Bug: 1030029
Change-Id: I078f63a77a2f06f720e15f139c2a81367ea114fe
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1947535
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>