DevTools: Add Sources setting option: Allow/Disallow scrolling past end of file
NOTE: this is a reupload of CL 1838270, targeting the new separate
DevTools repository.

https://chromium-review.googlesource.com/c/chromium/src/+/1838270

This PR addresses customer feedback received regarding scrolling
behavior for files viewed in the Sources tab. Currently, users are able
to scroll beyond the end of file, for all files at least two lines or
more in length. However, for extra-small files, this behavior is awkward
as the entire file fits inside the editor and scrolling is not needed.

As it turns out, this behavior is a feature. It is also present in
some other text editors, like VSCode. The purpose of scrolling past EOF
is to enable developers to scroll lines to the top of the file, even if
those lines are near the end of the file. After an informal discussion
with my peers, some prefer this feature, while others find it
unintuitive and distracting (myself included).

This change adds a new checkbox toggle option to the Sources setting:
"Allow scrolling past end of file"

When enabled (by default), the behavior is the same as it is now: users
may scroll beyond the end of file for all files >2 lines in length.

When disabled, the user may only scroll to the last line in the file,
and a scrollbar does NOT appear if the entire file fits inside the
editor. (When the editor is resized, a scrollbar appears as needed).

Animated GIF demo: https://i.imgur.com/zJytuf1.gif

Remark: the scroll past EoF feature is not native to CodeMirror and is
a Chromium specific customization for the Sources tab.

Bug: 998270
Change-Id: I79d11cf68c105aabe407269819f0495adfba44b7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1872878
Reviewed-by: Fabio Rocha <fabio.rocha@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Commit-Queue: Brian Cui <brcui@microsoft.com>