Generate proper tsbuildinfo and copy output files
The tsbuildinfo was being generated in the wrong location, which meant
that removing the generated folder would not properly retrigger the
typescript compiler. We now generate the tsbuildinfo next to the
tsconfig.json to make sure this works as expected.

Then we copy the files we have generated from ts_library into
out/<NAME>/resources/inspector so that the front_end can actually load
the files in the pak file.

Bug: 1011811
Change-Id: I20b76b6b41ce563a485bb34019d222da49cf4a7f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2013140
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>