Implement a Language-Specific Plugin System
This CL adds a plugin system to the debugger for language-specific
location mapping. Currently, a plugin supporting Wasm-DWARF for C/C++ is
included which is a proxy for an out-of-process backend. The plugin
system is currently guared behind an experiment.

Tests are currently missing because karma-testing is still blocked on a
mockable debugger, and layout testing is blocked because of the backend
requirement.


Bug: chromium:1042636
Change-Id: I1bf42afdf3688e67847e03acb027dd259403b6cc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2011851
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>