Introduce more accurate breakpoint decoration waiter for web tests
The usual way for breakpoint tests is to:
  1. Instrument the DebuggerPlugin
  2. Perfrom a breakpoint action (setting, toggling, etc.)
  3. Wait for a callback to be called
  4. Check CodeMirror if the breakpoint markers are set

In the case of inline or multiple breakpoints, the callback of step 3
is called multiple times but the test continues after the first time
the callback is called. This means not all decoration updates have
necessarily gone through.

This CL adds a more robust helper where we can pass in the expected
number of breakpoint decorations per line after the action is
performed. In case not all decoration updates have gone through,
we queue another round of waiting for decoration updates.

In essence, we turn step 3 above into a loop.

R=sigurds@chromium.org

Change-Id: I363123d3588e6f6878a82394438854ec4cacbff7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2027427
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>