Make audits_worker a worker with module scope
This allows us to properly import the Runtime with an ES import, as well
as resolve the service dispatcher in the appropriate scope. It relies on
the previous patch to be pushed to the appspot server, before it can be
resolved properly.

Bug: 1006759
Change-Id: I2d65c82e00ffaaa6ca2dd39fd4309e44d7a87002
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2030781
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>