Fix Karma typescript config to use root tsconfig.json
Karma typescript was using its own tsconfig, which became confusing as
soon as you were editing files that we already compiled with the root
tsconfig. This CL aligns the two, by specifying the `tsconfig` in the
karmaTypescriptConfig and removing all of the values that are not
strictly necessary.

Bug: 1009443
Change-Id: I905dc52a57e78391a6bca17454abeb775112d922
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2047193
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>