[utilities] Replace {keysArray} with {keys} plus spread operator
This CL is part of the cleanup effort to get rid of prototype
wrangling in utilities.js.

Note that {keysArray} on Set/Map can't be removed yet, as there are
a few use-sites in upstream chromium web tests that need to be cleaned
up first.

Bug: chromium:1050549
Change-Id: I3d5cdc9211763f53d0d43a110a67bba9485d0a11
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2046555
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>