Add e2etest for breakpoints in sources panel
The test will navigate through various part of the UI and open a file in
the sources panel. It will add a breakpoint by clicking on the
respective element in CodeMirror and verify that the top call-frame is
correct and has the correct url.

This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/sources/debugger/script-formatter-breakpoints-1.js;drc=ee7ea7dcf43928ef64337d8857205fa4bb4b4c5f?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest.

Bug: 1044632
Change-Id: I0de893dc679c723773e437c598d422def162526b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2047194
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>