High contrast work on Sources debugger panel
The changes fixes the hover and and keyboard focus on the panel. The
header gets a focus ring and when hovering or focusing the content
item, the background and foreground changes.

Before: https://imgur.com/a/gZD8xLg
After: https://imgur.com/a/LVruJ6P

Bug: 1048378
Change-Id: If697639f8a5b7d73844e1aab63c9bb2db884b295
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2055546
Commit-Queue: Kham Udom <khamu@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>