DevTools: scroll console scopes if more exist than can be seen
The scope menu is restricted to a specific height. However if a user
was to have more than would fit in that, they were unable to reach them.

This patch applies automatic overflow on the y axis to the menu.
If the amount of scopes fits in that height, they will have no scrollbar.
A scrollbar will appear when more than can fit within the menu
are present.

This allows users to get access to all scopes available regardless
of the menu height.

Screenshots:
[Before] - https://drive.google.com/file/d/1QK6n3HLUFPKPKRmKMp_K0z-s7E7fmOKc/view?usp=sharing
[After] - https://drive.google.com/file/d/16pFNHUe5oCNfasdpPcw_a5PSZrHFrRIX/view?usp=sharing

Bug: 1038461

Change-Id: Ia36b00d0b991a9614c1c7a2a2cf6e95540ca75e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2061128
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>