Add test for utf-8 character encoding
It also updates the hosted_mode server to serve JavaScript files with
the correct encoding. If we don't, the unicode characters that we use in
the DevTools implementations are garbled in either the source code or
test code.

This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/resource-tree/main-resource-content.js;l=1;drc=77578ccb4082ae20a9326d9e673225f1189ebb63?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest

Bug: 1044632
Change-Id: Ie6b714f33685bc0d67b63b24c61911b0c8ecefe3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2062588
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>