Add e2etest for cacheable responses in the network panel
I also discovered that the checkboxes we had in DevTools where not
properly accessible. While there was some logic to let a label element
be the aria-describedby for the input element, because they were both in
the shadow dom, it didn't do anything. Setting the appropriate
aria-label fixes that issue.

Secondly, the hosted_mode server now implements caching based on ETags.

This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/resource-tree/resource-tree-mimetype.js;l=1;drc=77578ccb4082ae20a9326d9e673225f1189ebb63?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest

Bug: 1044632
Change-Id: I3283b791d1d9a9673363d7cbb897d3f5c6c03ed0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2062603
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>