[utilities] Use Map + Set instead of Object + Array in ExtensionServer
This CL is part of the cleanup effort to get rid of Array#remove.
The ExtensionServer manages subscribers by using an object where
property names signify message types and the property itself is a list
of subscribers for that type. This can be better achieved using a Map
of Sets.

R=mathias@chromium.org

Bug: chromium:1050549
Change-Id: I1d642fb02f1e3bd0837c5df7c04343498810d3c6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2081827
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>