Concatenate all common ts_library into a single target
While performing the typescriptification process, we put all files into
separate ts_libraries. However, this is quite expensive and creates a
bottleneck for a chain of targets to compile.

To improve build performance, we can have 1 ts_library per module, which
should mediate that situation.

In the process, TypeScript started complaining about the definition of
`ls` having 2 arguments instead of 1. Initially I thought having 2
arguments is fine, but now it was required for TypeScript. This might be
a TS compiler bug in combination with declaration files, but it is
easier to simply remove the unused argument for now.

R=jacktfranklin@chromium.org

Bug: 1011811
Change-Id: Ifa29f89e800467fec55a2436db48bc0e4fb66089
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2089896
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>