[utilities] Replace Array#remove with helper in AnimationTimeline
This CL is part of a cleanup effort to get rid of Array#remove. In the
case of AnimationTimeline we can't use a Set as indices and indexed
access are used for the variable in question. Instead we are using
a helper version of the same code.

R=jacktfranklin@chromium.org

Bug: chromium:1050549
Change-Id: I3d0c880981acfdfad5671dac08f80422aed72bb5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2094125
Auto-Submit: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>