Fix 'Add custom device' focus regression
This recent CL [1] fixed an issue with default focus in the settings
screen by focusing the first tab header on launch, but this overruled
the default focus of the devices settings tab. The device settings tab
has an entry point in the device mode view (screenshot [2]) which is
intended to focus the 'Add custom device' button in the settings pane
when clicked. To fix this, I limit the tab header-focusing logic to only
run when the settings screen is opened via an action (i.e.
shortcut/button) and not when it is opening to reveal a specific view,
like it does for the devices settings.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2044982
[2] https://gyazo.com/782647efdb59f111614a900b9595e9d6

Bug: 1052248
Change-Id: I073a53aa9358685a842992c20745ae520507bdc0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2063057
Commit-Queue: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Reviewed-by: Shane Clifford <shanejc@microsoft.com>