Allow third_party DEPs to be exempt from import rules
We want to use the same rules for importing from third_party but
ultimately if the dep isn't quite structured in a way that lets us do
that we need to be able to make it exempt, which is what this CL does,
starting with `lit-html`.

We don't import `lit-html` on master yet but we do in the Breadcrumbs CL
[1] and we're unable to import everything from the root `lit-html.js`
file as it doesn't expose everything we need.

[1]:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2067066

Change-Id: I0678d5ee75e8dcb376cc40df6f9b37f32fbb557d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2097983
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>