Fix typeRoots inclusion for non-testonly ts_libraries
Prior to this change, TypeScript would not complain if you added the
following to (for example) common/ParsedURL.js:

import * as path from 'path';

That's because ts_library was eagerly including all
`node_modules/@types` in the typeRoots for any tsc invocation. However,
for non-testonly libraries (e.g. those targets in front_end) this is
invalid.

Therefore, if the ts_library is marked as test_only, it should
force-empty the typeRoots. I have confirmed that forcing the typeRoots
to an empty array ensures that no types are resolved, thus that
node_modules/@types is not automatically included.

R=jacktfranklin@chromium.org
DISABLE_THIRD_PARTY_CHECK=Update to TypeScript configuration

Bug: 1011811
Change-Id: I5aa9fd84cd8f674d4ede4e740e5cd0de58ff1e61
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2100913
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>