[profiler] Use global DebuggerWorkspaceBinding in legacy JS profiler
Currently we run into a "is not a function" error when clicking on
JavaScript in the flame graph. The ProfileView needs to use a concrete
instance of the DebuggerWorkspaceBinding, in this case the global one.

R=tvanderlippe@chromium.org

Change-Id: Id89cf1474e73f214192aafda1aa97a13bb607e35
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2100914
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>