Add SDK import that broke the media tab.
Is there a way to prevent this from happening with the presubmit checks?
This is the second time in a month that there are crashes here caused by
project-wide renaming / refactoring CLs. I only find out about these
issues because there are multiple users that open bugs on me.

Bug: 1054925
Bug: 1058320
Change-Id: I4fadef4203eb06de7095c0221cdef5726a2bb1ee
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2103702
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>