Fix command menu items for snippets
The runtime was throwing errors because `Snippets.SnippetsQuickOpen` did
not exist. The culprit is that the export statement still included a
`default class` rather than a `class`.

Also added an e2e-test that goes through the workflow of creating a
snippet and then verifying that the command menu shows these snippets
in its autocompletion.

R=aerotwist@chromium.org

Bug: 1060565
Change-Id: I9bb9d4ed34e5fda25cdd796a6a1e4c8a61f89958
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2107222
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>