Add affected resources for cookies to the Issues
This CL is crafted in a way to work with and without the back-end
changes, which are delayed because of the current situation. This
CL gets us unblocked in the front-end, so we can continue to iterate
and test the front-end implementation independent of progress in
the back-end.


Screenshot: https://imgur.com/a/eV9V0t4
Change-Id: I13fe4a02f0dc6af95e37bc8530019afced8449c0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2050230
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>