Rename check-type to check-type-closure
To reduce confusion as to which typechecker is running when executing
`npm run check-type`, rename this command to `check-type-closure`.
For those that invoke the previous `check-type` command, it will now
provide a descriptive message on which command to run for the
respective typechecker.

R=yangguo@chromium.org

Bug: 1011811
Change-Id: I1e8c1d64de2a3d3799e36ef8bc42dce76af44402
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2112369
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>