Prevent text truncation in default resolution
'Restore defaults and reload Dev-Tools' button of 'Dev-Tools' window
is seen chopped on increasing 'Minimum font size' to the 'maximum'.

The DevTools window can be resized to make it visible, but in the default
state, its truncated. This change restores the button text to
what it was previously.

Bug: 1064169
Change-Id: I051716bb64d1c9250196ae003697f98c184ea0b3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2117364
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Commit-Queue: Mike Jackson <mjackson@microsoft.com>