Fix size calculations in ExpandableTextPropertyValue
Previously, the code calculated the size of the string when using UTF-16
as the encoding. However, UTF-8 is much more common on the web and for
web development, so it makes more sense to use that. This CL changes
the size calculations to use UTF-8 instead.

Bug: 1024721
Change-Id: Iddc5950ffeccb2b561cd5301fb0aba3ec66a873f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2134287
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Surma <surma@chromium.org>