Add infobar about issues panel to console
This CL adds a sticky infobar to the console. It shows up whenever
issues are detected and notifies users about the existence of the new
issues panel. The bar disappears automatically when the user navigates
to a page for which no issues are detected.

Screenshot: https://imgur.com/Jk95x7n

Fixed: chromium:1061876
Change-Id: I9ec5aaab510021a77721a2b39b9e1b52197f4abb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2128132
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>