Remove explicit dependency on handlebars
We don't explicitly depend on handlebars in any of our scripts.
Instead, it is a sub-dependency for one of our dependencies.

To make sure we don't pin to an old version of Handlebars,
remove the explicit dependency for now.

R=jacktfranklin@chromium.org

Bug: 1068145
Change-Id: I63b7ef882b7a23c52e17a3e5bc5c1e22b5300e00
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2139975
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>