Fix mac-only enter key regression
This CL [1] changed the way certain actions are handled, including
sources.rename, which is bound to the enter key on mac. An unintended
consequence of this is that ShortcutRegistry now calls event.consume()
on all enter keydown events, preventing them from activating buttons.
The solution is to mark those actions as disabled when they are
registered and exclude them from ShortcutRegistry's search for
applicable actions.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2121565

Bug: 1070143
Change-Id: I6e73a00412f9df6163c0f587b66e2488ba8ac4f0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2147716
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>
Reviewed-by: Leo Lee <leolee@microsoft.com>
Commit-Queue: Jack Lynch <jalyn@microsoft.com>