Issues affected by network requests delay requestId to request mapping
Instead of trying to resolve a NetworkRequest object as soon as we
receive a new issue, we delay the resolve step when we render an
issue. This removes a timing dependency, otherwise the network request
must be present before the issue is reported to the frontend, which
is not always the case.

A follow up CL will improve the data structure in the network log,
to reduce the overhead of the requestId -> NetworkRequest step.

Drive-by: Redirect requests cause multiple NetworkRequest objects
with the same requestId. We display all requests with a given ID
since we don't know which specific one did have an issue.
E.g. cookies can be set on the inital redirect response (302) as
well as the final 200 response.

R=sigurds@chromium.org

Bug: chromium:1070577
Change-Id: Ia4b02347be9783a87c85e4477083c5314db71e1d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2150446
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>