Provide a message when contrast info is unavailable
We can't accurately gauge the background color in all situations due to
the possibility of multiple backgrounds, images, alpha channels and the
like. The code that calculates the contrast info is best effort, and so
always showing the contrast info would be less good when it's
potentially incorrect. This CL makes it clear that contrast information
is unavailable so that the developer knows and isn't guessing why it
shows on some occasions and not others.

A further improvement here would be to look at the C++ side and see if
there are ways to increase the amount of cases where we know the
background color underlying the text, but I'm not clear on how much
effort would be involved there.

See: https://imgur.com/a/YGu9ovg

Bug: 1033515,956711
Change-Id: If9ce252e7228cacf64b39bc45f50799a2f4d111e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2150457
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>