Devtools bug fix: Pretty print info bar
After clicking the 'Pretty print' button on the info bar
a new formatted file appears and the info bar is shown again.

This happens because a verification for formattable code is missing.
That is the same logic used by the button in the toolbar, so in this CL I'm adding that check for the infobar as well.

Before:
https://imgur.com/a/Cu2tkFp

After:
https://imgur.com/a/GaU1QHH

Bug: 1072251
Change-Id: I4be317ef0f8ba0321e31d29a070380d6e6018953
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2156054
Commit-Queue: Jose Leal <joselea@microsoft.com>
Reviewed-by: Brian Cui <brcui@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>