Make editable list items keyboard accessible
This CL adds tabindex to editable list items allowing selecting
list items using keyboard even if they have no focusable elements
inside. The user needs to be able to select the list item in order
to reveal the edit and delete buttons. Also, it sets the cursor:pointer
for toolbar buttons (used in the ListWidget and elsewhere) to
indicate that the buttons are clickable.

Bug: 1069471
Change-Id: I6141a263ea42ec3140157b929e5265da4853b951
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2159511
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>