Update README regarding rolling of dependencies
This section was previously hidden in the collapse details of the
second workflow. However, it is related to both the first and the
second workflow and therefore should be moved out.

Additionally, explicitly call out that it requires one of the first
two workflows to work, given its dependency on `gclient sync` as
well as the requirement for the DevTools infrastructure versioning
(which is different from the infrastructure as part of Chromium).

R=brgoddar@microsoft.com

Change-Id: I060351116ae79fe35f075f17de043762f486ccfd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2162889
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>