[issues] Report issue number instead of aggregated issue number
This change reports the number of issues known to DevTools instead of
the number of aggregated issues. This means that any improvement will
result in a decreased number. We think this provides better feedback for
developers.

To make this work properly, this change also introduces de-duplication
of issues, as cookie issues are ofen reported more than once by the
chromium back-end.

Note that until the fix in https://crrev.com/c/2164625 is in,
will possibly be too large when first opening DevTools. Reloading
the inspected page should work-around that bug.

Bug: chromium:1057979
Change-Id: Iebea90f3840a5c782866713990b03ae47712346b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2164746
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>