Remove long task candystriping from non-main thread tasks
The previous implementation of the candystriping coupled with the color
of the tasks. Since main thread tasks and worker tasks share the same
color this erroneously applied the candystriping for long tasks in
workers. This CL changes the implementation to inspect the level of the
bar in question, and only applies the candystriping when the level
matches that of the main thread.

Bug: 1075916
Change-Id: Iee57963c81d91a47676839a779cc3bb8c7741dbb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2174423
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>