Continue with typechecking NetworkManager with TypeScript
Issues discovered:
- Introduced CONNECTION_TYPES as a Map, as we were using it as an
object before
- Introduce MIME_TYPE enum to specify all mimetypes and strongly
type the getter and setter in NetworkRequest. Sadly we use strings
in other places, so we have to cast these :(
- It turns out that Common.ResourceType.resourceTypes was not
in sync with the protocol mime types. Most notably, it was missing
the `Ping` and `CSPViolationReport` keys. Additionally, it erroneously
includes `SourceMapScript` and `SourceMapStyleSheet` which do not
exist in the protocol. I have kept them in for now, as there is
other code relying on it. However, we should look into removing
them later if they indeed turn out to be obsolete.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Bug: 1011811
Change-Id: I67b4bba47317fdfab3d724b913465e1be416c6d6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2173370
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>