[issues] Only link cookie to network panel if a request is present
There are issues triggered from document.cookie in JavaScript, and it
doesn't make sense to link to the network request panel for such cookie
issues. We currently don't have the source location in the issue (this
requires back-end changes), so the option that should cause the least
confusion is to not link at all.

In some cases, the cookie might be available in the cookies section of
the application panel, but it is difficult to link there.

Bug: chromium:1057979
Change-Id: Id2eb0721397aa7c0e39cfcca2b829a6f57dcc4b3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2182417
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>