Move wasm_source_map to ES module; update Rust+wasm tools
This is attempt to replace
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2208981
with fixes in the Rust source.

As such, it also converts wasm_source_map to ES modules to match
direction of the rest of DevTools codebase and updates the way typings
are handled with a custom ad-hoc to please Closure Compiler.

Bug: chromium:1082720
Change-Id: I028ba5d759ec2a812a663923be107d52d13e2dfb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2208141
Commit-Queue: Ingvar Stepanyan <rreverser@google.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>