Auto resume all targets reloaded after crash
Since DevTools front-end is auto pausing all targets on start
it is responsible for resuming them. One of the scenarios
when a target has to be resumed is when it's reloaded after
crash. This is the case e.g. for service workers whose life
cycle looks like this:
1) Target.targetCreated (a new version of sw is created and started)
2) Inspector.targetCrashed (when the worker is stopped)
3) Inspector.targetReloadedAfterCrash (when the sw restarts)
4) Target.targetDestroyed (when the version of sw is stopped and replaced by a new one)
Events 2 and 3 are optional and may repeat multiple times if the
worker is stopped/restarted.

In such cases we don't want to show the crash dialog but are still
obliged to resume the target.

Bug: 1085150
Change-Id: Idb36b50bbeb09632cd8f1480d2f36b24387b7f2e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2213550
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yury Semikhatsky <yurys@chromium.org>