add style tag's media query rule to media query inspector
Currently the media query inspector does not show media queries created
by a style tag's "media=" attribute. This is because it's hard to
locate the origin of this media query within some stylesheet, because
it originates from an HTML document, not within a stylesheet. This CL
relaxes this prerequisite for the media query inspector and will list
non-locate-able media queries in the media query inspector. When trying
to locate the origin of this media query rule, it will show "empty",
which is the current behavior for existing cases of non-locate-able
media queries for different reasons.

Bug: chromium:1069467
Change-Id: Iff63af8317d0ad182dd5e6cfb1eb57ddf74ec8c4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2213429
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>