Add unused target flag to run_test_suite.py
We need to update the CQ bots to set the `target` flag for e2e-tests
to Release. However, we have to add a dummy flag first to change
the configuration. In a follow-up CL, we will actually start
using this flag in the e2e-test run.

R=aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1088463
Change-Id: Iada85d812e1a55b8095f2c4adabfb3a0985877d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2224814
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>