[wasm] Proper progress bar for Wasm disassembly generation
Change the wasmparser worker to emit progress events for the name
section decoding, the actual disassembly generation and the joining
of the lines. This way the progress bar in the front-end becomes
more useful for bigger wasm modules, where it's otherwise unclear
to the developer whether DevTools is still doing something or got
stuck somewhere due to a bug.

Fixed: chromium:1090262
Bug: chromium:1043030, chromium:10714320
Change-Id: I38403b17eb27da554855e9cc40e2d6d383df5e4d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2230535
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>