[Performance] Bumps reload wait time
When performing the reload and record action in the timeline we wait 3
seconds after the load event fires before stopping the trace. This tends
to mean that we miss the InteractiveTime event from Blink, which in turn
means we don't get any Total Blocking Time values.

This CL pushes back the wait time to 5 seconds to increase the
probability that we will get TBT from Blink. In the longer term this
will need to be superseded by listening for the InteractiveTime event
directly rather than the Load event.

Change-Id: I8ee53ddb118f9b5db0ed2cc2a11ee5ae82cae1e6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2238231
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Auto-Submit: Paul Lewis <aerotwist@chromium.org>