Split computed-pane-properties_tests.ts to avoid timeouts
This test has been failing in CI environments because of timeout issues.
Because this test was doing 2 things: testing a property on one node and
then another property on another node, it's simple enough to split this
in 2 test cases, and therefore avoid exceeding the per-test-case
timeout.

Bug: 1096439
Change-Id: Ie7c94ad9db4a34d86dbc3f443e68611739b0531b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2251364
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jose Leal <joselea@microsoft.com>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>