[browser_sdk] Refactor main frame navigated event listening
This CL introduces a TopFrameNavigated event on the FrameManager.
Unfortunately, this event is not useful to front_end/sdk, as our
layering currently forbids importing browser_sdk there.

However, other places in devtools may use this event to listen
for navigations on the main frame.

Note that in contexts are related to the state of V8, the right
event to listen to might be executionContextDestroyed or even
executionContextsCleared.

Bug: chromium:1069378
Change-Id: Ib56d51c9a0d27346ef91e81029d49ae4c2c66fff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2270167
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>