Fix formatting conflict with async test
This CL fixes an issue we have where ESLint and Clang are clashing. We enforce
via ESLint that there is a space after an `async` keyword, e.g:

```
async () => // good
async() => // bad
```

But in the `async` that has changed in this CL Clang was formatting it
incorrectly. ESLint would then fix it, but then Clang would undo it, and so on.

Because these functions don't need the `async` keyword, we can drop it,
avoiding the problem. I will follow this up with a small reproduction case and
report a bug in Clang - I suspect the TS defs alongside the `async` func are
throwing it off slightly.

CRBug for clang issue: https://bugs.chromium.org/p/chromium/issues/detail?id=1100841

Change-Id: I0eb79ce36861cb8307800832d090ae77eb25191d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2274506
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>