Run terser minifier as part of rollup_entrypoint
We are not running the minifier as part of build_release_applications.py,
as the build performance would be significantly worse. We don't pay that
high of a price with `rollup_entrypoint`, as Ninja will not rebuild if
it is not necessary.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Fixed: 1084578
Change-Id: Id18fb826aef33c9efddfe390a62eec69485c6295
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2256989
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>