Remove legacy support that may crash the DevTools renderer
The DevTools renderer crashes if it receives a request to record
histogram data {sample, boundary} where sample is not a non-negative
integer smaller than boundary, or where boundary is larger than the
actual boundary of the histogram in the back-end.

The first call that gets removed falls in to the latter category,
the boundary 100 is way too large. The second call is likely to
trigger the former category, as we currently have ~40 panel codes,
and the hard-coded boundary is 20.

Bug: chromium:1098736
Change-Id: If9c2f13978c130aa8ff5ebe7b421da0c3026c838
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2274511
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>