tests: Make e2e --test-file arg relative to gen dir and document it
Previously the --test-file arg was relative the root of the project
which meant it was easy to accidentally specify the .js file in the
source tree rather than the one in out/Default/gen/.

This CL changes the argument to be a relative path from gen/test/e2e
so that the user can just write the name of the test file, which is
always going to be under e2e/ for e2e tests.

We always need to run the .js file, but users might type the .ts
file out of habit, so accept .js, .ts as well as no extension, as
long as we can find the file.

Also add documentation to e2e/README.md that explains how to use the
argument.

Bug: chromium:1056140
Change-Id: I0c55768d8541072c31e011427393dc8d2aa3580b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2278472
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>