Fix snippet saving and search
Snippet names were escaped using the built-in escape function which
was not working well for certain characters. This CL replaces the
usage of the escape/unescape functions with encodeURIComponent/
decodeURIComponent. Also it fixes the search in snippets where
the snippet file URL was not correct.

Bug: chromium:1066415
Change-Id: Ic690a0574c087310c9e58746161e0c74d12ae8c1
Fixed: 1043121
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2279844
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Eric Leese <leese@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>