Generate tsconfig in both debug and release build
Previously, we were generating the tsconfig only in release builds.
However, debug builds require the same treatment. Therefore,
refactor `devtools_entrypoint` to generate the tsconfig in both
cases.

We need to separate out the copying of the declaration file, as that
still only happens when we run Rollup in a release build.

This is a temporary fix to unbreak master and the inclusion of the
first custom element. We will clean up this logic in a follow-up CL.

R=aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1098730, 1101738, 1061037
Change-Id: I231d87b6788b76c88028a80c27dd742e04dc70de
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2282518
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>