List formatter_worker_entrypoint and acorn.mjs in GRD
Neither formatter_worker_entrypoint.js nor acorn.mjs were listed in the
GRD for DevTools. This meant that when someone went to pretty print file
contents the runtime 404'd on the two files in Release mode, and pretty
printing failed.

This CL adds them to the GRD inputs so that they are bundled in Release

TBR=jacktfranklin@chromium.org

Bug: 1107859
Change-Id: Icec02ad4b0fc024de25a4d3b5621a35f586a2890
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2316072
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>