Clarify missing shortcuts for screen reader users
This CL adds an aria-label for rows in the custom shortcuts settings tab
without shortcuts, e.g. "Step" in this screenshot [1] would be read by a
screenreader as "Step No shortcut for action" rather than just "Step".
The purpose of this label is to make it clearer to screen reader users
when an action doesn't have any shortcuts, as we received some feedback
that the existing experience is confusing.

[1] https://i.imgur.com/0rP7EaP.png

Bug: 963183
Change-Id: I6178d56399487dfa08117f1a408ae79faff3fc0b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2321473
Reviewed-by: Shane Clifford <shanejc@microsoft.com>
Commit-Queue: Jack Lynch <jalyn@microsoft.com>