Adding a link to SW internals page to view other origin SWs in DevTools
This is a follow-up CL to removing the UI for removing the drop down
menu to view service worker registrations from other origins. Instead,
this CL includes a link to chrome://serviceworker-internals that opens
in a new tab.

I will be adding a UMA counter to see how many users use the link with
the 'devtools' query parameter.

Bug: 807440
Change-Id: I55571134989f4e3b381cf61840f8090337f7b1e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2324268
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@google.com>