Source order viewer toggle persistence
Currently, when navigating through the element tree with the source
order viewer (SOV) on, it will be toggled off when an element with no child elements is encountered (since it has no source order to show).

This CL changes it so that the SOV will stay active even if these types of elements are encountered.

Before/after gifs: https://imgur.com/a/tjqVy36

Change-Id: I42838091e0eb9bf0d7394f0b211db7b92ebaa2b7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2335876
Commit-Queue: Olivia Flynn <t-olfly@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>