[issues] Filter third-party issues in IssuesManager
This CL adds a new 'isCausedByThirdParty' getter to the 'Issue' class.
Third-party issues are considered out of the control of the developer,
so we will hide them by default.

This CL introduces a new setting, that can be toogled by the user via
some yet-to-be-determined UX and causes the Issues tab to either
show or hide third-party issues.

This filtering happens on the IssuesManager. All issues are stored,
but only Issues that are accepted by the filter will cause events
to be fired and are returned by 'IssuesManager#issues()'.

This change has currently no impact, as all Issues are marked
as first-party issues by default.

R=sigurds@chromium.org

Bug: chromium:1080589
Change-Id: I6f3254d0f4655e170c4385c205092e6f7097b3f9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2339135
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>