Tidy up type reference look-ups in bridges
This CL moves the logic for looking up a reference to one place in the component
bridges. This is needed as soon we're going to support enums and we need to be
able to easily look up a string reference and find the right interface/type
alias/enum.

Change-Id: I108d84be8343989b0b94e51431bf79af4a7f7afb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2339555
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>