Remove the Service Worker from appearing if origin changes in DevTools
Currently, if the user navigates to a page like youtube.com which has
a Service Worker, and then navigates to a page with no Service Worker
such as NewTab, the youtube.com registration still appears in the
main panel of the Service Worker tab in the Application Panel in
DevTools. The correct behaviour for this situation would be that if
the origin changes, the service worker should no longer appear in the
main panel, and only when viewing all other origins. This CL aims
to fix this issue.

Bug: 807440
Change-Id: I514da5fea40cf2505ba093728d17c760234c42c0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2351792
Commit-Queue: Nidhi Jaju <nidhijaju@google.com>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>