Change 'View Trace' button to indicate throttling method
Lighthouse issue:
https://github.com/GoogleChrome/lighthouse/issues/9971

The 'View Trace' button can lead to confusing results when using
simulated throttling because the timings in the original trace do not
match the metric results.

This CL renames the 'View Trace' button to be 'View Original Trace' and
adds a tooltip to clarify the situation.  If active throttling is used
then the button reverts to 'View Trace' and hides the tooltip.

Change-Id: I8ade46c3031dafc381237559dde27181529177fb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2339861
Commit-Queue: Adam Raine <asraine@google.com>
Reviewed-by: Connor Clark <cjamcl@chromium.org>