Expose isInDarkMode test from get-stylesheet
This CL exposes a `isInDarkMode()` test to components. The use case
is for components to get `ui/inspectorSyntaxHighlight.css` with dark
mode support. This css file is special because it's dark mode is
written separately in `ui/inspectorSyntaxHighlightDark.css`, so the
components need to determine when to call
getStyleSheets(`ui/inspectorSyntaxHighlight.css`) in light mode, and
getStyleSheets(`ui/inspectorSyntaxHighlightDark.css`) in dark mode.

Change-Id: I5a6b79b6c884874a74bf6321431c82d22ddf4875
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2362570
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>