Performance: Fix perf issue with Event Log
This CL fixes a performance issue that occurs in the event log.
Attempting to select the first node on a tree refresh (that occurrs after
toggling one of the checkboxes) was creating a selection event
that caused a repeated chain of rebuilding the event tree.
Suppressing the selection event still produces the desired behavior:
the first element is selected when the tree is rebuilt after
changing the event listener filters, but no longer enters a massive
recursive loop.

Before is in bug
After: https://imgur.com/NUrTwUj

Bug: 1112663
Change-Id: Ica51890eeb23234dcb60d59ebfc2815165d16979
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2358069
Reviewed-by: Jose Leal <joselea@microsoft.com>
Reviewed-by: Junyi Xiao <juxiao@microsoft.com>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>