Fix bug in async function _pollComputedStyleUpdates
The async function _pollComputedStyleUpdates checked whether
_isCSSPropertyTrackingEnabled before doing an await, but not again
after that. The state can change during the await. This lead to
an object becomming null in the meantime, leading to an exception.

Additionally, the dispose() of the CSSModel would trigger a
message to the agent of a already closed target.

Bug: chromium:1066772, chromium:1121534
No-Tree-Checks: true
Change-Id: Ife63bd3a68d6440ffbb2113c78204bae2cf81d43
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2374209
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>