Grid: Match the label colors with the line colors
In the grid overlay it is possible to configure the color of the row and
column lines. This will be useful when we'll make it possible for users
to change those colors to make sure the overlay is always visible
against a variety of backgrounds.

This change makes it so that the background color of the grid labels
match the chosen line colors.
The get the background color, we blend the line color with a 0.2 opacity
white.

This change also picks the text color for the labels from white or black
depending on which one has the best color contrasta with the background.

Here are a few examples: https://imgur.com/a/BbvrqrF

Bug: 1121152
Change-Id: I3ba4341b9d8674244ca3dd32f3512c304ae6dda3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2372742
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jose Leal <joselea@microsoft.com>